requirements.txt
# Direct dependencies for avalon_mm_width_adapter test runner
# Install with:  pip install -r requirements.txt
#
# Backend / LLM backends (all use the same openai-compatible client):
#   Ollama (local, free):    ollama pull llama3.2  →  AI_BASE_URL=http://localhost:11434/v1
#   GitHub Models (free):   AI_BASE_URL=https://models.inference.ai.azure.com  AI_API_KEY=<PAT>
#   Groq (free tier):       AI_BASE_URL=https://api.groq.com/openai/v1  AI_API_KEY=<KEY>

# Web UI
Flask==3.1.3

# AI chat (OpenAI-compatible client, works with Ollama / GitHub Models / Groq)
openai==2.30.0

# VUnit simulation framework
vunit_hdl==4.7.1
